Call methods sans performSelector; fix setDebugMode.
Adds an optional protocol. This, ideally, allows the compiler to operate
slightly closer to what seems correct since it now has method signatures
to work with (regardless of whether the object actually implements
the protocol).

setDebugMode now passes an actual BOOL, as expected by the plugin
protocol. Previously broken because the C++ side would pass an NSNumber
when all plugins are written to take a BOOL, so the BOOL would be
truncated and this would determine its truthiness. As such, it's much
simpler to just pass a BOOL.

Closes #36.